Add .flutterflowignore support
This file allows for users to create a list of files to be ignored
when exporting the project code (works a bit like .gitignore).

Files matching the globbing rules listed at .flutterflowignore are
going to be ignored. This might break the project if changes in these
files are needed to compile the project, so it should be used with
extreme caution.

This functionality was previously achieved with:
https://github.com/krabhishek/flutterflow-filtered-pull